modules: split report
